Changed inheritance hierarchy to remove some redundancy; unnest some inner classes for clarity
