+++ src/ispd/motor/carga/TaskBuilder.java
@@ -3,7 +3,6 @@ package ispd.motor.carga;
 import ispd.motor.filas.RedeDeFilas;
 import ispd.motor.filas.Tarefa;
 import ispd.motor.filas.servidores.CS_Processamento;
-import ispd.motor.random.Distribution;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -12,13 +11,6 @@ import java.util.stream.Stream;
 
 public abstract class TaskBuilder {
     static final double FILE_RECEIVE_TIME = 0.0009765625;
-    protected final Generator<Integer> idGenerator;
-    protected final Distribution random;
-
-    protected TaskBuilder() {
-        this.idGenerator = new SequentialIntegerGenerator();
-        this.random = new Distribution(System.currentTimeMillis());
-    }
 
     public List<Tarefa> makeTasksEvenlyDistributedBetweenMasters(
             final RedeDeFilas qn, final int taskCount) {
